content/canvas/test/test_2d.canvas.reference.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.canvas.reference</title>
<!-- Testing: CanvasRenderingContext2D.canvas refers back to its canvas -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(ctx.canvas === canvas, "ctx.canvas === canvas");

SimpleTest.finish();

});
</script>
